requirements.txt
tensorflow==2.20.0
numpy==2.3.0
matplotlib==3.10.0
pillow==11.0.0
streamlit==1.41.0
scikit-learn==1.7.0
opencv-python==4.8.1.78
jupyterlab==4.3.0
ipython==8.30.0
